requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
dominate==2.6.0
firebase-admin==5.2.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.4
proto-plus==1.20.6
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.19.2
PyYAML==6.0
requests==2.28.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.27
svgwrite==1.4.1
uritemplate==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
